# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Photo Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List photos for the current user
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: archived
          in: query
          schema: { type: boolean }
        - name: favorited
          in: query
          schema: { type: boolean }
        - name: albumId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    post:
      operationId: createPhoto
      tags: [photos]
      summary: Register a photo after upload to object storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Photo created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /photos/stats:
    get:
      operationId: getPhotoStats
      tags: [photos]
      summary: Dashboard statistics for the current user
      responses:
        "200":
          description: Photo statistics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PhotoStats" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /photos/timeline:
    get:
      operationId: getPhotoTimeline
      tags: [photos]
      summary: Photos grouped by date for timeline view
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200":
          description: Timeline groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /photos/{id}:
    get:
      operationId: getPhoto
      tags: [photos]
      summary: Get a single photo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Photo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    patch:
      operationId: updatePhoto
      tags: [photos]
      summary: Update photo metadata
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUpdate"
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Move photo to trash
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Moved to trash
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /photos/{id}/tags:
    post:
      operationId: addPhotoTag
      tags: [photos]
      summary: Add a tag to a photo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagInput" }
      responses:
        "200":
          description: Updated photo with tags
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /photos/{id}/tags/{tag}:
    delete:
      operationId: removePhotoTag
      tags: [photos]
      summary: Remove a tag from a photo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: tag
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated photo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ─── Albums ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Photos ──────────────────────────────────────────────────────────────
    Photo:
      type: object
      required:
        - id
        - userId
        - filename
        - objectPath
        - mimeType
        - fileSize
        - width
        - height
        - isFavorited
        - isArchived
        - isTrashed
        - tags
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        userId: { type: string }
        filename: { type: string }
        objectPath: { type: string }
        thumbnailPath: { type: ["string", "null"] }
        mimeType: { type: string }
        fileSize: { type: integer }
        width: { type: integer }
        height: { type: integer }
        isFavorited: { type: boolean }
        isArchived: { type: boolean }
        isTrashed: { type: boolean }
        trashedAt: { type: ["string", "null"], format: date-time }
        caption: { type: ["string", "null"] }
        aiTags: { type: ["string", "null"] }
        exifData: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        albums: { type: array, items: { $ref: "#/components/schemas/AlbumRef" } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PhotoInput:
      type: object
      required: [filename, objectPath, mimeType, fileSize, width, height]
      properties:
        filename: { type: string, minLength: 1 }
        objectPath: { type: string, minLength: 1 }
        thumbnailPath: { type: string }
        mimeType: { type: string, minLength: 1 }
        fileSize: { type: integer, minimum: 1 }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        caption: { type: string }
        exifData: { type: string }
    PhotoUpdate:
      type: object
      properties:
        filename: { type: string, minLength: 1 }
        isFavorited: { type: boolean }
        isArchived: { type: boolean }
        caption: { type: string }
    PhotoList:
      type: object
      required: [photos, total, page, limit]
      properties:
        photos:
          type: array
          items: { $ref: "#/components/schemas/Photo" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    PhotoStats:
      type: object
      required:
        - totalPhotos
        - totalFavorited
        - totalArchived
        - totalInTrash
        - totalStorageBytes
        - totalAlbums
        - recentUploads
        - uploadsByMonth
      properties:
        totalPhotos: { type: integer }
        totalFavorited: { type: integer }
        totalArchived: { type: integer }
        totalInTrash: { type: integer }
        totalStorageBytes: { type: integer }
        totalAlbums: { type: integer }
        recentUploads:
          type: array
          items: { $ref: "#/components/schemas/Photo" }
        uploadsByMonth:
          type: array
          items: { $ref: "#/components/schemas/MonthCount" }
    MonthCount:
      type: object
      required: [month, count]
      properties:
        month: { type: string }
        count: { type: integer }
    TimelineGroup:
      type: object
      required: [date, photos]
      properties:
        date: { type: string }
        photos:
          type: array
          items: { $ref: "#/components/schemas/Photo" }
    TimelineResponse:
      type: object
      required: [groups, total]
      properties:
        groups:
          type: array
          items: { $ref: "#/components/schemas/TimelineGroup" }
        total: { type: integer }
    TagInput:
      type: object
      required: [tag]
      properties:
        tag: { type: string, minLength: 1 }

    # ─── Albums ──────────────────────────────────────────────────────────────
    AlbumRef:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...