# lib/api-zod/src/generated/types/photoStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Photo Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthCount } from './monthCount';
import type { Photo } from './photo';

export interface PhotoStats {
  totalPhotos: number;
  totalFavorited: number;
  totalArchived: number;
  totalInTrash: number;
  totalStorageBytes: number;
  totalAlbums: number;
  recentUploads: Photo[];
  uploadsByMonth: MonthCount[];
}