# lib/api-zod/src/generated/types/timelineGroup.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Photo Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from './photo';

export interface TimelineGroup {
  date: string;
  photos: Photo[];
}