# lib/api-zod/src/generated/types/timelineResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Photo Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineGroup } from './timelineGroup';

export interface TimelineResponse {
  groups: TimelineGroup[];
  total: number;
}